Bug fixes for pack/unpack of eps.
Now copying all data from eps.el appropriately (much of it is for output,
but may be important). Changed from memcpy of damage object to packing of
individual variables. Copying the full structure makes a copy of function
pointers which may be inconsistent accross processes. Furthermore, the
function pointers are automatically reset upon unpacking of the data
similarly to the cohesive elements. In order to maintain consistency, i.e.
be able to diff memory/restart files, the function pointers are no longer
copied.

It turns out that this bug was very minor, and did not lead to incorrect results
however, fixing it was necissary to ensure correctness and keep future devs
from falling down the same rabbit hole.