Properly set adaptive rebalancing heuristic.
Heuristic is based on maximum server time (i.e. estimated WC time for
all RUCs on all servers to be finished computing). Note that this is
really the most important measure in terms of deciding whether to
migrate cells as we are most concerned w/ faster time to solution.
Faster simulation = lower CPU hours = better comp. efficiency. The
server wait time is a good metric of utilization, but overall time is
what you pay for.