bug fix: multiscale restart
Found and fixed a bug with multiscale restart in which microscale
servers were not dumping the correct state for the cell.

Details:
When a solution was computed with more than one cell assigned to each
server, restarting from any point in the history (particularly after
initiation of inelastic processes) yielded incorrect results. It was
discovered that only the first solution on each server was dumped to
restart files FOR ALL cells on the server. As a result, the same
microscale state was used to reinitialize all cells with varied
results.