full implementation of new constitutive modeling interface with access to hyperelastic material response.
The newer data structure separates the state variables and the model
parameters/functions. The interface for the hyperelastic response is
implemented in plasticity_model_none.c. Use this implementation as a
reference for new models. This implementation compiles, but has not
been tested for functionality as it requires implementation of the
overall data structure throughout the rest of the code.