Implement the BPA model equations
All of the BPA model equations are implemented with a _private_
interface accessible through the `_plasticity_model_BPA.h` _private_
header file. This header file should only be included in the BPA
implemenataion and testing drivers. This file may be removed in the
future and all of the functions declared static.

This commit also contains the initialization of the integration
algorithm and full implementations of main interface to retun elastic
stresses/tangents.