conditionally use Lagrange multiplier in BPA staggered solution.
If the constraint is identically satisfied, then the system of
equations including the Lagrange multiplier becomes indeterminate. In
this case, we only solve the subset of the system of equations that
are not satisfied (i.e., the M/Wp blocks). The 'solve_Ax_b' routine
has been modified to enable using only the sub-matrix and sub-residual
in the solution process.