add some const declarations to the crystal plasticity implementation. Minor formatting updates
*Important Programming Practice*
Using the const declarations not only aid the compiler in performing
optimizations but enable others to better understand what is input and
output from a function when there is little/no documentation. This is
particularly important when things are passed by reference as changes
can be propagated out of the function through the arguments in this
way.