change definition of plasticity_ctx. Elastic deformation computed internally from current (n + 1) state.
This required a change to libCM as well, please see the
'extend-material' branch. There are now additional user-defined
functions called from libCM that update the internal state of the
Constitutive_model object. In propagating this interface, there seems
to be slight change in the convergence (some steps in my test example
require an additional PDE NR iteration), however the solution is the
same within 10^-7 in L2.