add restart read/write functions and some get_F functions for IVD
Implemented the read/write functions for restart via the CM interface
implementation. There is an open question in terms of whether or not
integer flags (e.g., damaged) should be included in the restart data.

In addition, the functions for getting values at n-1 have been
removed. Thus, there will not be support for iniertial in the initial
implementation of the model. This can be added back in, but may
require more thought.

issue #37