add IVD driver program to test the CM interface
The driver computes a simple approx. uniaxial load history and outputs
the time, sigma_zz and damage parameter to a file `ivd.dat`. If the
file `ivd.props` exists, it is read for the damage parameters. Note
that a valid line (non-blank/comment) is required after specification
of the damage parameters to avoid an abort on EOF.

*NOTE* The driver name for the BPA model was changed, it is now simply
 `BPA_driver`

issue #37