fix minor bug in VTK output when configured with --no-vtk
When building the code without linking against the VTK libraries, the
macroscopic displacements that were output for visualization were
computed for interface deformations regardless if bulk MS modeling was
used. Note that this *did not* effect the actual computation of
fields, only the output macro displacements for visualization. If the
code was configured to link against the VTK libraries, then the output
displacements were correct.