add elems with fixed nodes to list of elems with prescribed displacements
Think of fixed nodes as nodes with prescribed disp u=0. This is
important for correctly computing the reactions. However, there is a
slight performance cost in that for each new loading step, we are
unnecissarily computing zeros. Note that this cost is small in
comparison to the remainder of the computation in that the boundary is
typically << than the volume.

issue #41