implement function to compute reaction forces associated with each pre disp
The new function computes the reaction force corresponding to each
magnitude of prescribed displacement to `PGFEM_stdout`. The 1st entry
is the resultant force for all fixed nodes *in all directions*. If one
wishes to print the resultant forces in each direction for fixed
nodes, then the user must prescribe multiple zero-valued
displacements. The sum of all printed values should equal zero up to
the magnitude of ||R||.

issue #41