make the maximum damage value (`ome_max` in [0, 1)) an input model parameter through the CM interface.
The model parameter specification for the `ISO_VISCOUS_DAMAGE` and
`J2_PLASTICITY_DAMAGE` models now include a maximum value for the
damage state variable. This is specified after the damage viscosity in
the `model_params.in` file (please see the
(wiki)[https://gitlab-cswarm.crc.nd.edu/pgfem_3d/pgfem_3d/wikis/mult-mat-cm]
for more details).

Note that if the user specifies `ome_max >= 1`, the implementation
automatically resets it to 0.9999.