implement `-disp-cm` functionality for `-SS` analysis
CM models may now be used in single-scale analysis through the
displacement formulation (used routines called in `-disp` analysis).
The results are verified for the hyper-elastic and J2/damage
models. Note that these are the only models currently supported as
they do not rely on the multiplicative decomposition in the FE
implementation. Other models (e.g., BPA or crystal plasticity) my not
converge and/or may give incorrect results.

issue #38