reset CM state vars after computing the reactions
Computing the reactions calls the functions for computing residuals on
elements. This may in turn call the CM integration algorithm. For
rate-dependent constitutive models, this may disturb the integration
of the model by allowing it to relax. Therefore, we reset the state
variables to remove this fictitious relaxation from the load history.