Brought back the restrict keyword.
"restrict" is not a C++ keyword, so we must set up whater is supported
in a given compiler.  We do this during configuration, and the results
are stored in the "config.h" file. Before this commit, "config.h" was
not included everywhere.  I included it in "PGFEM_mpi.h", and I made
sure that this file is included everywhere.  This may or may not be
correct.  If "PGFEM_mpi.h" is supposed to be a part of a portable
reusable library distributed to users, then it should not include
"config.h".