Update the configuration to support c++14.
This adds the autoconf-archive macros to test the C++ compiler to make sure that
it supports C++14, as well as to add any flags that are necessary to do so.

This also consolidates some of the logic for checking to see if the compiler
supports MPI when the user wants MPI.

This closes issue #65.