More autotools modernization.
This commit restructures the configuration and automake files to clean up the
formatting and eliminate some cruft. It should not make any functionality
changes, other than the elimination of use of the "special linker" flag from the
autotools build. I can not currently find the purpose of this and rewriting the
LINKER rules isn't necessarily the correct way to handle this.

This closes issue #73.