Remove `__cplusplus` checks.
During conversion from C to C++ these checks were added to support the
multi-language nature of the project. The project is now entirely C++14 and
these checks are no longer appropriate.

A side effect of this is that it unmaskes the symbol collisions that occur with
the Generalizsed_constitutive_model math helper functions that were originally
copied from PGFEM. These should be dealt with in the Gcm project itself by
updating the visibility of those symbols to be hidden. For now, I updated the
Makefile.am in `src` to allow multiple definitions.