Update the Matrix pseduo-template to be an actual template.
The correct option here would be to just update PGFEM to use a real linear
algebra template library like MTL (or TTL if it works in this case). Replacing
the existing macro definitions with templated C++ code is the second best
option.

This addresses many of the -fpermissive warnings that are generated because of
the type-unsafety of the previous code.