Remove unused `j` variables and provide initial values for `f2` and `AL`.
The `f2` and `AL` values could be read before being initialized in the first
iteration of the while loop if the else branch is taken. My assumption is that
that branch is guaranteed to take the first arm in the first iteration, but I'm
not sure. I updated the declarations of these two variables to initialize them
to the default value for a `double`.