Use a typed pointer to represent the fast maps.
The void* that was being used is both unnecessary and causes complications with
the C++ compiler. Forward declaring the map class is the appropriate way to do
this. In the future it would make more sense to just have map values in the
communicator, rather than pointers, as they are 1) small, 2) not polymorphic,
and 3) important for performance.