Add functionality to verify the return value from fscanf.
Often, system headers have fscanf defined with no-ignore-return attributes. This
is important because the API from fscanf says that it should return the actual
number of fields that were matched, and the user has to verify that this matches
the expected number of fields.

This adds a variadic macro function to check this return value, and applies it
where warnings are occurring.