Clean up use of matrix templates.
During refactoring some of the matrix templates changed slightly from their
previous macro counterparts. This commit cleans up some point-of-use changes
required to use the macros. This mostly consists of allocating matrix handles
with calloc instead of malloc, to make sure that the structures are
0-initialized.

This also fixes a bug in the Matrix template where the default copy constructor
was being used rather than the user-defined copy constructor.