Update our configuration of vtk, and update the --enable-vtk default to yes.
Rather than asking the user to provide include and library paths, along with all
of the libraries that PGFem3D wants (which is a complicated problem for the
user), we just do the standard option for asking for the installation root of
the vtk installation.

The complexity is that vtk is

1. Installed by default relative to its version number, so we can't find it
without that.
2. Extremely dependent on its version, which makes it hard to know what
libraries actually exist.
3. Built with hundreds to thousands of libraries.

We deal with this as best we can, by providing the option to pass a version
number in addition to an install root, and looking for libraries that we think
we know are necessary.