Work out the bugs with the crystal_plasticity make check target.
It turns out that there is an input file directory in the verification_scripts
version of this test that is required in order to get the expected results. I do
not know how to specify the path to these files inside of PGFem3D, so instead I
copy the files into the build tree in order to run the tests.