Fix allocation error discovered by gcc 7.1.
The integer value `-1` was being returned from functions that get the
workspace's number of send or receive partners in certain cases. This value was
then being passed as a `size_t` into an allocation function. Unfortunately, `-1`
is a very large `size_t` and would generate issues. I adjusted this so that `0`
is returned in these corner cases rather than `MAX_SIZE_T`.