Begin solver abstraction layer.
This sets up the build infrastructure for a `solvers` convenience library in the
source, and moves the current `SOLVER_OPTIONS` into a new Solvers.h header with
a new name. We add a `using SOLVER_OPTIONS = pgfem3d::Solver` definition in the
existing header to limit the exposure of this change for now.