Replace the legacy API for the hypre solver with member functions.
This creates a SparseSystem abstraction that encapsulates

1. The data structures containing the coefficient data for the sparse matrix.
2. The partitioning of the data.
3. The preconditioner.
4. The implict solver.

After this commit all of the details and includes that touch Hypre are contained
behind this API. The client code should not care that Hypre is being used at
all. This commit also performs multiple style-based patches to legacy code that
was encountered during this process.