Add a couple of rpaths for vtk and suitesparse to the build.
Libtool doesn't interact well with the outside world, and the cswarm system has
an even older libtool version than we normally use, so it may have bugs in
it. This commit forces some rpaths into the executable link lines so that they
find the right versions of vtk and suitsparse.