Stop allowing multiply defined symbols.
The generalized constitutive model (GCM) repository defines many of the same
math symbols that PGFEM does. We have been dealing with these name collisions by
asking the linker toi use the first instance of the symbol that it encounters.

We've upgraded the GCM repository to build a libtool library rather than a
simple .a, and---for whatever reason---libtool seems able to deal with this
collision without any additional flags.

This commit eliminates the extra linker flag that we were using, and fixes a
header-defined function that needs to be marked `static inline`.