Use 0-based indexing for TTL.
This entire file seems to be using non-standard [1,n] indices. C/C++ code should
always use [0,n) unless forced to use 1-based indices by a library out of our
control. Even in that case, the loops should be written as [0,n) and the index
should be adjusted to the non-standard index at the point of use.