Move an allreduce out of a `catch` clause.
Collective communication in a `catch` will only work if all ranks take the same
exception. This `catch` is related to initializing comm hints, and the allreduce
is trying to see if anyone had an error. If only one rank has an error then the
code wouldn't terminate.

This performs the allreduce after the local `catch`.