Add an implementation of a blocked row distribution.
PGFem3D's global sparse system is distributed in a blocked manner, but code to
deal with this blocking is replicated and distributed throughout the code
base. This is the first step in centralizing that code.

This introduces a simple blocked row distribution template that takes an array
of row counts and establishes a row-rank mapping. It provides simplified
interfaces to test if a rank owns a row, find which rank owns a row, find the
minimum and maximum number of rows that a rank owns, and compute the number of
rows that a rank owns.

This first commit simply generates a distribution during the `Psparse_ApAi`
initialization and checks to make sure that its implementation agrees with the
historical C-based implementations for the same functionality. Future commits
will incrementally replace the C-based implementations until we are only left
using the C++ template interface.