Use the BlockedRowDistribution to avoid an N^2 loop in `Psparse_ApAi`.
There is some code where the `Psparse_ApAi` routine counts the number of degrees
of freedom that it's going to have to send to each rank, and initializes the
`SLID` with the local IDs that will be sent. It was being initialized using a
loop that looked like:

```
for each rank
  for each non-local id
    if rank owns id
      rank count ++
```

This uses the owner function to convert that outer loop from the number of
processes to log2 of the number of processes.

```
for each non-local id
  owner id count ++
```