Add degree-of-freedom enumerators to the element and cohesive element classes.
These classes help to define PGFem3D's sparse graph structure. Each element
contains a list of a number of different degrees of freedom (ids) and we use
these lists to build up the sparse system in Psparse_ApAi.cc.

This commit will begin the modernization of this construction algorithm. Each
element "knows" its degrees of freedom, and how and where they are stored. We
provide encapsulation of this information using `forEach` enumerators that can
apply a lambda function to each degree of freedom id.