Add a utility class that supports RAII for FILE* operations.
This is part of modernization of the code base. We rely on fprintf throughout
the code and this will allow us to properly handly closing files in the face of
errors without a lot of extra nonsense.

The recommended usage is:

```
if (auto file = scoped_fopen(path, mode)) {
  CHECK_SCANF(file, ...);
  fprintf(file, ...);
}
```

The RAII mechanism will close the file at the end of its scope, as if it is a
C++ file stream, and it will also close the file properly if anything in the
`if` clause (like the CHECK_SCANF) throws an exception.