Normalize use of `const` and `struct` to define class types.
These identifiers are not equivalent in C++ and if a type is defined as a
`class` then it needs to be declared as a `class` everywhere, while if a type
is defined as a `struct` then it needs to be declared as a `struct` everywhere.

Ideally we wouldn't have any `struct`s however this isn't feasible given the
legacy C nature of the code base.